Rename the zod-compiled case to zod (compiled)
