+++ cases/index.ts
@@ -58,7 +58,6 @@ export const cases = [
   'vality',
   'yup',
   'zod',
-  'zod-compiled',
   'zod3',
   'deepkit',
   'effect-schema',
